backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
pymysql
google-generativeai
pillow
python-dotenv
cryptography
requests
